feat: Add function to read website YAML file
The code changes add a new function `findWebsiteYAML` that recursively searches for the `website.yml` file in the given directory and its parent directories. It returns the content of the YAML file if found, or `null` otherwise. The function is used in the `readWebsiteYAML` function to read the `website.yml` file in the workspace folder.

This commit message suggests adding the function to read the website YAML file.