feat: Modify getComponentInfo to return componentInfo and entityName
The code changes in PowerPagesChatParticipantUtils modify the getComponentInfo function to return an object with two properties: componentInfo and entityName. This change improves the handling of component information and entity details for the active file. The updated function is used in PowerPagesChatParticipant to retrieve the necessary data for sending API requests.