[PowerPages] Copilot Availability Check for @powerpages (#962)
* feat: Add environmentID field to OrgDetails interface

The code changes add an optional `environmentID` field to the `OrgDetails` interface in the `PowerPagesChatParticipantTypes.ts` file. This field represents the ID of the environment associated with the organization. It is used in the `getEndpoint` function in the `PowerPagesChatParticipantUtils.ts` file to retrieve the intelligence endpoint and geo name for the specified organization and environment.

Co-authored-by: amitjoshi <amitjoshi@microsoft.com>
Co-authored-by: tyaginidhi <tyaginidhi@users.noreply.github.com>

* feat: Update PowerPagesChatParticipantUtils to return componentInfo and entityName

The code changes in PowerPagesChatParticipantUtils modify the getComponentInfo function to return an object with two properties: componentInfo and entityName. This change allows for better handling of component information and entity details for the active file. The updated function is used in PowerPagesChatParticipant to retrieve the necessary data for sending API requests.

Co-authored-by: amitjoshi <amitjoshi@microsoft.com>

* feat: Modify getComponentInfo to return componentInfo and entityName

The code changes in PowerPagesChatParticipantUtils modify the getComponentInfo function to return an object with two properties: componentInfo and entityName. This change improves the handling of component information and entity details for the active file. The updated function is used in PowerPagesChatParticipant to retrieve the necessary data for sending API requests.

---------

Co-authored-by: amitjoshi <amitjoshi@microsoft.com>