feat: Add site creation inputs for Power Pages
The code changes in PowerPagesChatParticipantUtils add a new function called getSiteCreationInputs. This function prompts the user to enter the site name, environment name, and domain name for creating a new Power Pages site. It uses the MultiStepInput utility to guide the user through the input process. The collected inputs are returned as an object.

Co-authored-by: amitjoshi <amitjoshi@microsoft.com>