[@powerpages] Chat Variable and Org Handling (#1029)
* remove variables

* Updated org auth handling

* refactor: Renamed fxn

* refactor: Initialize organization details in OrgHandlerUtils

This commit refactors the `initializeOrgDetails` function in `OrgHandlerUtils.ts` to improve the initialization of organization details. It now retrieves stored organization details from the global state and returns them if available. If not, it uses the PAC wrapper to fetch the organization details and stores them in the global state for future use. This change enhances the efficiency and reliability of the organization details initialization process.

---------

Co-authored-by: amitjoshi <amitjoshi@microsoft.com>